Allow diagram to be saved as SVG image
Lots of tweaks and workarounds were needed to make it compatible with
typical SVG software like cairo, eog, imagemagick, or inkscape.  The
biggest issue has to do with styles.

  - Support for <style/> with classes is very poor among the
    viewers/editors.  Therefore, embedding a stylesheet directly within
    <style/> was scrapped.
  - Embedding inline styles is not exactly trivial:
      - You can't just naively use getComputedStyle because (1) this can
        cause misstyling because sometimes you do want things to "fall
        through" and inherit from the parents (most notably, we
        intentionally omit styling on symbols so they can be overridden)
        (2) the output is also extremely bloated because it includes
        styles from all its ancestors.
      - We have to use the 'matches' API, but that requires having a
        list of CSS selectors that we can match against, i.e. we need to
        read and parse the CSS.

To avoid ajax (for loading CSS, which doesn't work on file:// due to the
same-origin policy) and to avoid parsing the CSS at run time, webpack +
postcss-js was used, which required redoing much of the build system
(not that there was much of that anyway).  postcss-js doesn't have a
built-in "syntax" for outputing in raw css-js format, so a custom
wrapper (stringify.js) was used.  It had to be in a separate file too
because the stable version of postcss-loader doesn't yet support
lambdas.

There were also some other more minor tweaks that had to be done:

  - 'transparent' and 'rgba' don't work; need to use *-opacity
  - <symbol/> defined outside <defs/> causes inkscape to crash
    when moving an element with the cursor
  - zero radius in arcs is not supported in inkscape (causes glitchy
    rectangles to appear)
  - commas between transforms is not supported in inkscape nor eog
    (have to use spaces)
  - href is not supported in many viewers/editors; have to use
    xlink:href instead (even though the latter is deprecated)